refactored Categories & added method to Models for deleted items
Categories are now normal classes (with a label as the only attribute)
and are managed by the ToDoItemModel (as they are to simple to create
an extra model for them). All models got updated for that.

In addition all models return now with getToDoItems() only undeleted
items. All deleted items can get retrieved with getDeletedToDoItems()